Se agrega lógica para la venta .

+++ ServicioGnc/Models/DetalleVenta.cs
@@ -8,7 +8,7 @@ namespace ServicioGnc.Models
         public int DetalleVentaId { get; set; }
         public Nullable<int> ProductoId { get; set; }
         public Nullable<int> VentaId { get; set; }
-        public Nullable<int> Cantidad { get; set; }
+        public Nullable<double> Cantidad { get; set; }
         public Nullable<double> Precio { get; set; }
         public Nullable<double> Subtotal { get; set; }
         public virtual Producto Producto { get; set; }
